order class volume type change

+++ include/Order.hpp
@@ -22,7 +22,7 @@ public:
 
 private:
   float price;
-  size_t volume;
+  int volume;
 };
